Refactor upload service and types
This commit refactors the upload service in the admin app to use the new TFile type instead of TImage. It also updates the deleteFiles method to accept an array of TFile objects instead of public_ids. The TImage type is renamed to TFile in the types/image/index.ts file.